+++ src/ZoomedWaveform.cpp
@@ -52,7 +52,7 @@ void ZoomedWaveform::paint(juce::Graphics& g)
 
 		for (auto i = 0; i < cueTargets.size(); ++i) {
 			if ((cueTargets[i]->first * audioThumb.getTotalLength()) > left && (cueTargets[i]->first * audioThumb.getTotalLength()) < right) {
-				g.setColour(juce::Colour::fromHSL(cueTargets[i]->second, 1, 0.5, 1));
+				g.setColour(juce::Colour::fromHSL(static_cast<float>(cueTargets[i]->second), 1.0f, 0.5f, 1.0f));
 				double widthPos = juce::jmap(cueTargets[i]->first * audioThumb.getTotalLength(), left, right, (double)0, (double)getWidth());
 				g.drawRect(widthPos, 0, 1, getHeight());
 			}
